requirements.txt
google-generativeai
python-dotenv
tk